refactor: Move StatVarsMap to its own file
Moved the StatVarsMap class from stat_var_processor.py to its own file, stat_vars_map.py, to improve modularity.

This change also includes:
- Adding a new test file for the StatVarsMap class.
- Fixing DuplicateFlagError and NameError issues that arose from the move.
