+++ tools/statvar_importer/mcf_diff.py
@@ -74,44 +74,55 @@ sys.path.append(
 from counters import Counters
 from mcf_file_util import load_mcf_nodes, filter_mcf_nodes, normalize_mcf_node, normalize_value, node_dict_to_text, get_node_dcid, strip_namespace
 
-flags.DEFINE_string('mcf1', '', 'MCF file with nodes')
-flags.DEFINE_string('mcf2', '', 'MCF file with nodes')
-flags.DEFINE_list(
-    'ignore_property',
-    [
-        'description',
-        'provenance',
-        'memberOf',
-        'member',
-        'name',
-        'constraintProperties',
-        'keyString',  # 'Node'
-        'relevantVariable',
-    ],
-    'List of properties to be ignored in diffs.',
-)
-flags.DEFINE_list('compare_property', [],
-                  'List of properties to be compared in diffs.')
-flags.DEFINE_bool(
-    'fingerprint_dcid',
-    False,
-    'If set, ignores the dcid for nodes and instead uses fingerprint.',
-)
-flags.DEFINE_list(
-    'ignore_nodes_with_pv',
-    [],
-    'Ignore nodes containing any of the property:values in the comma separated'
-    ' list.',
-)
-flags.DEFINE_list('compare_dcids', [],
-                  'Compare nodes with dcids in the comma seperated list.')
-flags.DEFINE_list(
-    'compare_nodes_with_pv',
-    [],
-    'Compare nodes containing any of the property:values in the comma separated'
-    ' list.',
-)
-flags.DEFINE_bool('show_diff_nodes_only', True, 'Output nodes with diff only.')
+# TODO: Move the flags to the main function and pass them as arguments to the
+# library functions.
+if 'mcf1' not in flags.FLAGS:
+    flags.DEFINE_string('mcf1', '', 'MCF file with nodes')
+if 'mcf2' not in flags.FLAGS:
+    flags.DEFINE_string('mcf2', '', 'MCF file with nodes')
+if 'ignore_property' not in flags.FLAGS:
+    flags.DEFINE_list(
+        'ignore_property',
+        [
+            'description',
+            'provenance',
+            'memberOf',
+            'member',
+            'name',
+            'constraintProperties',
+            'keyString',  # 'Node'
+            'relevantVariable',
+        ],
+        'List of properties to be ignored in diffs.',
+    )
+if 'compare_property' not in flags.FLAGS:
+    flags.DEFINE_list('compare_property', [],
+                      'List of properties to be compared in diffs.')
+if 'fingerprint_dcid' not in flags.FLAGS:
+    flags.DEFINE_bool(
+        'fingerprint_dcid',
+        False,
+        'If set, ignores the dcid for nodes and instead uses fingerprint.',
+    )
+if 'ignore_nodes_with_pv' not in flags.FLAGS:
+    flags.DEFINE_list(
+        'ignore_nodes_with_pv',
+        [],
+        'Ignore nodes containing any of the property:values in the comma separated'
+        ' list.',
+    )
+if 'compare_dcids' not in flags.FLAGS:
+    flags.DEFINE_list('compare_dcids', [],
+                      'Compare nodes with dcids in the comma seperated list.')
+if 'compare_nodes_with_pv' not in flags.FLAGS:
+    flags.DEFINE_list(
+        'compare_nodes_with_pv',
+        [],
+        'Compare nodes containing any of the property:values in the comma separated'
+        ' list.',
+    )
+if 'show_diff_nodes_only' not in flags.FLAGS:
+    flags.DEFINE_bool('show_diff_nodes_only', True, 'Output nodes with diff only.')
 
 _FLAGS = flags.FLAGS
 # _FLAGS(sys.argv)  # Allow invocation without app.run()
