+++ tools/statvar_importer/mcf_file_util.py
@@ -53,15 +53,21 @@ from absl import logging
 
 _FLAGS = flags.FLAGS
 
-flags.DEFINE_string('input_mcf', '', 'List of MCF files to load.')
-flags.DEFINE_string('output_mcf', '', 'output MCF nodes loaded into file.')
-flags.DEFINE_bool(
-    'append_values',
-    True,
-    'Append new values to existing properties. If False, new values overwrite'
-    ' existing value.',
-)
-flags.DEFINE_bool('normalize', True, 'If True, values are normalized.')
+# TODO: Move the flags to the main function and pass them as arguments to the
+# library functions.
+if 'input_mcf' not in flags.FLAGS:
+    flags.DEFINE_string('input_mcf', '', 'List of MCF files to load.')
+if 'output_mcf' not in flags.FLAGS:
+    flags.DEFINE_string('output_mcf', '', 'output MCF nodes loaded into file.')
+if 'append_values' not in flags.FLAGS:
+    flags.DEFINE_bool(
+        'append_values',
+        True,
+        'Append new values to existing properties. If False, new values overwrite'
+        ' existing value.',
+    )
+if 'normalize' not in flags.FLAGS:
+    flags.DEFINE_bool('normalize', True, 'If True, values are normalized.')
 
 _SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(_SCRIPT_DIR)
