+++ tools/statvar_importer/stat_vars_map_test.py
@@ -1,0 +1,40 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Unit tests for stat_vars_map.py."""
+
+import os
+import sys
+import unittest
+
+from absl import app
+from absl import logging
+
+_SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(os.path.dirname(_SCRIPT_DIR))
+sys.path.append(os.path.dirname(os.path.dirname(_SCRIPT_DIR)))
+sys.path.append(
+    os.path.join(os.path.dirname(os.path.dirname(_SCRIPT_DIR)), "util"))
+
+from stat_vars_map import StatVarsMap
+
+
+class TestStatVarsMap(unittest.TestCase):
+
+    def test_create_stat_vars_map(self):
+        stat_vars_map = StatVarsMap()
+        self.assertIsNotNone(stat_vars_map)
+
+
+if __name__ == "__main__":
+    app.run(unittest.main)
